+++ Microsoft.Azure.Cosmos/src/Telemetry/Collector/TelemetryToServiceCollectorNoOp.cs
@@ -1,0 +1,35 @@
+﻿//------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+//------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.Telemetry.Collector
+{
+    using System;
+    using System.Net.Http;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Microsoft.Azure.Cosmos;
+    using Microsoft.Azure.Cosmos.Core.Trace;
+    using Microsoft.Azure.Cosmos.Handler;
+    using Microsoft.Azure.Cosmos.Query.Core.Monads;
+    using Microsoft.Azure.Cosmos.Resource.Settings;
+    using Microsoft.Azure.Cosmos.Routing;
+    using Microsoft.Azure.Cosmos.Telemetry;
+    using Microsoft.Azure.Cosmos.Tracing;
+    using Microsoft.Azure.Cosmos.Tracing.TraceData;
+    using Microsoft.Azure.Documents;
+    using Microsoft.Azure.Documents.Collections;
+
+    internal class TelemetryToServiceCollectorNoOp : IClientTelemetryCollectors
+    {
+        public void CollectCacheInfo(string cacheName, Func<TelemetryInformation> functionFordata)
+        {
+            //NoOps
+        }
+
+        public void CollectOperationAndNetworkInfo(Func<TelemetryInformation> functionFordata)
+        {
+            //NoOps
+        }
+    }
+}
